Adds an +initialize method to DSSparseArray to test if NSIndexSet has it enumerateIndexesUsingBlock method fixed.
This is a change from testing it lazily when the enumeration methods are called. With the new way (+initialize) it is always done, but only once, without having to test status every time an enumeration method is called. With the old way it might never be done if the user never uses an enumeration method. So, if enumeration methods are consistently and frequently used the new method is better. If the enumeration methods are infrequently used or only called once per application execution the old method is better.

The enumeration bug seems to be fixed in Xcode 6 released with Yosemite!